+++ core/app-version.js
@@ -1,1 +1,1 @@
-export const APP_VERSION = '20260701-1';
+export const APP_VERSION = '20260701-2';