+++ features/config-controller.js
@@ -1,4 +1,4 @@
-import { normalizeConfig } from '../runtime/config-service.js?v=20260701-1';
+import { normalizeConfig } from '../runtime/config-service.js?v=20260701-2';
 
 export class ConfigController {
     constructor(dom) {